split up observer files to separate data implementations
